Merged PR 33: The flow of Making a new Order/PreOrder is finished.
The flow of Making a new Order/PreOrder is finished.

* Changed DBController to be a singleton (like the others) and moved him into the controller folder(package)
* New class "Controllers" under controllers package will be responsible of making an instance of each of the controllers (will be called upon on server run).
  * by calling "import static controller.Controllers.*;" u will be able to call our controllers by just their name for example dbController.
* Added LastCarRemovalException (to be supported on server\gui) to notify the user he is trying to the delete the last car assigned to him.
* CustomerController will now use the DBController in order to insert the new orders into the DB AND get their UID.

Tests changes:
In order to support our current order tests and future tests of items that are handled by the DB, I added a isTest param to DBController which will be checked when calling upon its insertOrder method and return 1 as UID (instead of acctualy working with the server) this way we can tell that if the test fails it wasnt due to the server (mocking).
this will be needed to happen in each method used by DB and also used by the tests. for more info on how this works just look at orderTest.

bugs: atm inserting an order doesnt work on the DB is beeing checked by Or.