Merged PR 39: Added OrderStats(enum) and Multiple Subscriptions per carID is now supported.
Added OrderStats(enum) and Multiple Subscriptions per carID is now supported.

UPDATE : FIXED THE BUGS NOW CONTAINS A FULL WORKING (AND TESTED) FLOW WITH THE DATABASE ON CREATING AN ORDER!

nothing much to add to this, adjusted the code to work as we discussed on skype.
for more info check the changes themselves.

Also made sure DB will be instanced first on the CPS SERVER.