Merged PR 41: Added ParkingController
**Added ParkingController**

**ParkingController**
Added ParkingController which is a singleton, the class is responsible for:
Iniating the parking lot.
Entering a parking lot.
Exiting a parking lot.
And to do on next branch: To get the current status map of the parking lot and export it to a PDF file format.

**ParkingLot**
Changed ParkingLot class to have indicators of current slots being occupied in the 3 dimensions ( width/height/depth).
Added another constructor for DB purposes.