Merged PR 73: Merge OptimizeAndFixLoadoutFromDB to master
Optimized import from db of Orders and Subscriptions

Added an important todo....i am using the mapper with PutAll function (of his) and putting in <Integer,object> instead of first casting it into the right object. it might be dooing it on it's own but we need to test it....too much data currently inorder to test it properly i suggest dooing it after cleanup of DB
and if need to cast it by myself its a fast fix.