Merged PR 74: Merge MakeReports to master
**This is not for a Real Merge yet!**
 (will w8 for Or to merge his then Rebase and then check with Rami since he has the finish of this flow to see that it actualy works!)
This is just so that u can review the work so far and see if it's ok by u.

**Started Working on Making all of the Reports (in a generic way) WIP....**
For now:
public String makeReportFromDB(Report.ReportType reportType, Integer managerID)

will be a generic method to create ALL of the Reports accourding to their type. manageID == 999 is default to CPS Tasker.
There is a generic Template used in this method in order to make a nice Table on a simple String.
Return value can just be printed out or (WIP by Rami) printed as a PDF official report (to be on the System)

Currently supports: Making of : DAILY_FINISHED_ORDERS, DAILY_CANCELED_ORDERS