Merged PR 86: Added SQL sanitation to Strings
# Added SQL sanitation to Strings
 Added SQL sanitation through StringUtils. Use `StringUtils.sanitizeForSQL(text)` to get a sanitized string for
 insertion, and then use `StringUtils.desanitizeFromSQL(text)` to get a de-sanitized string.
I already went over 'dbController' and used it where I thought necessary, feel free to add more if you see the need.

# Added duplicate email verification before creating new users
In addition to limiting it on the SQL level, I added a simple check before creation that makes sure the email isn't already registered as a customer or employee. At the moment the handling of this error is NOT YET implemented, but it should be with @<abarda02@campus.haifa.ac.il> 's next PR.